+++ demo/Cimpress.Nancy.Demos/Cimpress.Nancy.Demos.xproj
@@ -7,7 +7,7 @@
   <Import Project="$(VSToolsPath)\DotNet\Microsoft.DotNet.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
     <ProjectGuid>4aca5cab-84a8-4267-ae9f-92737153fe63</ProjectGuid>
-    <RootNamespace>Cimpress.Nancy.Demo</RootNamespace>
+    <RootNamespace>Cimpress.Nancy.Demos</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">.\obj</BaseIntermediateOutputPath>
     <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
