+++ demo/Cimpress.Nancy.Demos/Program.cs
@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.IO;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Builder;
 
-namespace Cimpress.Nancy.Demo
+namespace Cimpress.Nancy.Demos
 {
     public class Program
     {
